Refactor Utils constants and fix UploadImageScreen
