+++ app/src/main/java/ar/edu/unlam/mobile/scaffolding/evolution/data/database/ConstantsAvatarUrl.kt
@@ -1,0 +1,10 @@
+package ar.edu.unlam.mobile.scaffolding.evolution.data.database
+
+object ConstantsAvatarUrl {
+    const val IMAGES = "userAvatarImages"
+    const val URL = "url"
+    const val CREATED_AT = "createdAt"
+    private val UID: String? = null
+    val IMAGE_NAME: String = "$UID.jpg"
+    const val ALL_IMAGES = "image/*"
+}
