+++ app/src/main/java/ar/edu/unlam/mobile/scaffolding/evolution/data/database/ConstantsUserData.kt
@@ -1,0 +1,11 @@
+package ar.edu.unlam.mobile.scaffolding.evolution.data.database
+
+object ConstantsUserData {
+    const val USERDATA = "UsersFutureFight"
+    const val EMAIL = "email"
+    const val NAME = "name"
+    const val NICKNAME = "nickname"
+    const val USERID = "userID"
+    const val AVATAR = "avatarUrl"
+    const val USERINFO = "infoUser"
+}
