--- app/src/main/java/ar/edu/unlam/mobile/scaffolding/evolution/data/network/utils/Constants.kt
@@ -1,26 +1,0 @@
-package ar.edu.unlam.mobile.scaffolding.evolution.data.network.utils
-
-object Constants {
-    const val IMAGES = "userAvatarImages"
-    const val USERDATA = "UsersFutureFight"
-    const val URL = "url"
-    const val CREATED_AT = "createdAt"
-    private val UID: String? = null
-    val IMAGE_NAME: String = "$UID.jpg"
-    const val ALL_IMAGES = "image/*"
-    const val EMAIL = "email"
-    const val NAME = "name"
-    const val NICKNAME = "nickname"
-    const val USERID = "userID"
-    const val AVATAR = "avatarUrl"
-    const val USERINFO = "infoUser"
-}
-
-// object Constants {
-//    const val IMAGES = "images"
-//    const val URL = "url"
-//    const val CREATED_AT = "createdAt"
-//    const val UID = "avatar"
-//    const val IMAGE_NAME = "$UID.jpg"
-//    const val ALL_IMAGES = "image/*"
-// }
